+++ package.json
@@ -25,6 +25,7 @@
     "dompurify": "^3.0.3",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
+    "framer-motion": "^10.12.17",
     "iron-session": "^6.3.1",
     "jwt-decode": "^3.1.2",
     "next": "^13.4.7",
