+++ src/app/layout.tsx
@@ -7,7 +7,7 @@ import { ConfigProvider, theme } from 'antd'
 
 const inter = Poppins({
   display: "swap",
-  weight: ["500", "600", "700"],
+  weight: ["600", "700"],
   subsets: ['latin']
 })
