+++ src/app/writer-dashboard/create-article/create-article.tsx
@@ -17,12 +17,12 @@ import type { UploadFile } from 'antd/es/upload/interface';
 import Image from "next/legacy/image";
 import { useSelector } from 'react-redux';
 import { GetCategoriesResponse, useCreateArticleMutation, useGetCategoriesMutation } from '@/services';
-import { TypedFormDataValue, TypedFormData, getTypedFormData } from '@/utils/formDataTyper';
+import { TypedFormDataValue, TypedFormData, getTypedFormData } from '@/utils';
 import WriterLayout from '@/layout/Head/Writer/WriterLayout';
 import Heads from '@/layout/Head/Head';
 import { TextEditor } from '@/components/TextEditor';
 import { RootState } from '@/store';
-import { DbConcurrencyError, ErrorResponse, InternalServerError } from '@/utils/errorResponseHandler';
+import { DbConcurrencyError, ErrorResponse, InternalServerError } from '@/utils';
 
 export interface ArticleData {
   body: string;
