+++ src/components/Button.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Button, ButtonProps } from 'antd';
 
-interface StandardButtonProps extends ButtonProps {
+interface Props extends ButtonProps {
   label?: string;
   round?: string;
   active?: string;
@@ -13,7 +13,7 @@ interface StandardButtonProps extends ButtonProps {
  * @param ... props from Ant Design <Button/> element, including `onClick()`
  * @returns {JSX.Element} button element
  */
-export default function StyledButton(props: StandardButtonProps): JSX.Element {
+export default function StyledButton(props: Props): JSX.Element {
   const { label, round, active } = props;
   const roundClass = round || 'rounded-full';
   const activeClass = active || '';
@@ -21,7 +21,7 @@ export default function StyledButton(props: StandardButtonProps): JSX.Element {
   return (
     <Button
       {...props}
-      className={`${props.className ?? ''} ${roundClass} ${activeClass}`}
+      className={`${props.className ?? ''} ${roundClass} ${activeClass} font-bold text-xl flex justify-center items-center px-4 py-5`}
     >
       {label}
     </Button>
