+++ src/utils/formatter.ts
@@ -11,3 +11,7 @@ export const formatCurrency = ({
   const formattedValue = formatter.format(Number(value));
   return symbol ? `Rp.${formattedValue}` : formattedValue;
 };
+
+export function dateFormat(date: string) {
+  return new Date(date).toLocaleDateString('id-ID', { weekday: "long", year: "numeric", month: "short", day: "numeric" })
+}
